(1) Add Trust to DeviceCryptoService name (2) Move DeviceTrustCryptoService under auth folder
