+++ libs/common/src/auth/services/device-trust-crypto.service.spec.ts
@@ -1,25 +1,25 @@
 import { mock, mockReset } from "jest-mock-extended";
 
-import { DevicesApiServiceAbstraction } from "../abstractions/devices/devices-api.service.abstraction";
-import { DeviceResponse } from "../abstractions/devices/responses/device.response";
-import { EncryptionType } from "../enums/encryption-type.enum";
-import { AppIdService } from "../platform/abstractions/app-id.service";
-import { CryptoFunctionService } from "../platform/abstractions/crypto-function.service";
-import { EncryptService } from "../platform/abstractions/encrypt.service";
-import { StateService } from "../platform/abstractions/state.service";
-import { EncString } from "../platform/models/domain/enc-string";
+import { DevicesApiServiceAbstraction } from "../../abstractions/devices/devices-api.service.abstraction";
+import { DeviceResponse } from "../../abstractions/devices/responses/device.response";
+import { EncryptionType } from "../../enums/encryption-type.enum";
+import { AppIdService } from "../../platform/abstractions/app-id.service";
+import { CryptoFunctionService } from "../../platform/abstractions/crypto-function.service";
+import { EncryptService } from "../../platform/abstractions/encrypt.service";
+import { StateService } from "../../platform/abstractions/state.service";
+import { EncString } from "../../platform/models/domain/enc-string";
 import {
   SymmetricCryptoKey,
   DeviceKey,
   UserKey,
-} from "../platform/models/domain/symmetric-crypto-key";
-import { CryptoService } from "../platform/services/crypto.service";
-import { CsprngArray } from "../types/csprng";
+} from "../../platform/models/domain/symmetric-crypto-key";
+import { CryptoService } from "../../platform/services/crypto.service";
+import { CsprngArray } from "../../types/csprng";
 
-import { DeviceCryptoService } from "./device-crypto.service.implementation";
+import { DeviceTrustCryptoService } from "./device-trust-crypto.service.implementation";
 
-describe("deviceCryptoService", () => {
-  let deviceCryptoService: DeviceCryptoService;
+describe("deviceTrustCryptoService", () => {
+  let deviceTrustCryptoService: DeviceTrustCryptoService;
 
   const cryptoFunctionService = mock<CryptoFunctionService>();
   const cryptoService = mock<CryptoService>();
@@ -35,7 +35,7 @@ describe("deviceCryptoService", () => {
     mockReset(appIdService);
     mockReset(devicesApiService);
 
-    deviceCryptoService = new DeviceCryptoService(
+    deviceTrustCryptoService = new DeviceTrustCryptoService(
       cryptoFunctionService,
       cryptoService,
       encryptService,
@@ -46,7 +46,7 @@ describe("deviceCryptoService", () => {
   });
 
   it("instantiates", () => {
-    expect(deviceCryptoService).not.toBeFalsy();
+    expect(deviceTrustCryptoService).not.toBeFalsy();
   });
 
   describe("Trusted Device Encryption", () => {
@@ -68,14 +68,14 @@ describe("deviceCryptoService", () => {
         ) as DeviceKey;
 
         stateSvcGetDeviceKeySpy = jest.spyOn(stateService, "getDeviceKey");
-        makeDeviceKeySpy = jest.spyOn(deviceCryptoService as any, "makeDeviceKey");
+        makeDeviceKeySpy = jest.spyOn(deviceTrustCryptoService as any, "makeDeviceKey");
       });
 
       it("gets a device key when there is not an existing device key", async () => {
         stateSvcGetDeviceKeySpy.mockResolvedValue(null);
         makeDeviceKeySpy.mockResolvedValue(mockDeviceKey);
 
-        const deviceKey = await deviceCryptoService.getDeviceKey();
+        const deviceKey = await deviceTrustCryptoService.getDeviceKey();
 
         expect(stateSvcGetDeviceKeySpy).toHaveBeenCalledTimes(1);
         expect(makeDeviceKeySpy).toHaveBeenCalledTimes(1);
@@ -88,7 +88,7 @@ describe("deviceCryptoService", () => {
       it("returns the existing device key without creating a new one when there is an existing device key", async () => {
         stateSvcGetDeviceKeySpy.mockResolvedValue(existingDeviceKey);
 
-        const deviceKey = await deviceCryptoService.getDeviceKey();
+        const deviceKey = await deviceTrustCryptoService.getDeviceKey();
 
         expect(stateSvcGetDeviceKeySpy).toHaveBeenCalledTimes(1);
         expect(makeDeviceKeySpy).not.toHaveBeenCalled();
@@ -111,7 +111,7 @@ describe("deviceCryptoService", () => {
 
         // TypeScript will allow calling private methods if the object is of type 'any'
         // This is a hacky workaround, but it allows for cleaner tests
-        const deviceKey = await (deviceCryptoService as any).makeDeviceKey();
+        const deviceKey = await (deviceTrustCryptoService as any).makeDeviceKey();
 
         expect(cryptoFuncSvcRandomBytesSpy).toHaveBeenCalledTimes(1);
         expect(cryptoFuncSvcRandomBytesSpy).toHaveBeenCalledWith(deviceKeyBytesLength);
@@ -191,7 +191,7 @@ describe("deviceCryptoService", () => {
 
         // TypeScript will allow calling private methods if the object is of type 'any'
         makeDeviceKeySpy = jest
-          .spyOn(deviceCryptoService as any, "makeDeviceKey")
+          .spyOn(deviceTrustCryptoService as any, "makeDeviceKey")
           .mockResolvedValue(mockDeviceKey);
 
         rsaGenerateKeyPairSpy = jest
@@ -227,7 +227,7 @@ describe("deviceCryptoService", () => {
       });
 
       it("calls the required methods with the correct arguments and returns a DeviceResponse", async () => {
-        const response = await deviceCryptoService.trustDevice();
+        const response = await deviceTrustCryptoService.trustDevice();
 
         expect(makeDeviceKeySpy).toHaveBeenCalledTimes(1);
         expect(rsaGenerateKeyPairSpy).toHaveBeenCalledTimes(1);
@@ -253,7 +253,7 @@ describe("deviceCryptoService", () => {
         // setup the spy to return null
         cryptoSvcGetUserKeyFromMemorySpy.mockResolvedValue(null);
         // check if the expected error is thrown
-        await expect(deviceCryptoService.trustDevice()).rejects.toThrow(
+        await expect(deviceTrustCryptoService.trustDevice()).rejects.toThrow(
           "User symmetric key not found"
         );
 
@@ -263,7 +263,7 @@ describe("deviceCryptoService", () => {
         // setup the spy to return undefined
         cryptoSvcGetUserKeyFromMemorySpy.mockResolvedValue(undefined);
         // check if the expected error is thrown
-        await expect(deviceCryptoService.trustDevice()).rejects.toThrow(
+        await expect(deviceTrustCryptoService.trustDevice()).rejects.toThrow(
           "User symmetric key not found"
         );
       });
@@ -303,7 +303,7 @@ describe("deviceCryptoService", () => {
           it(`throws an error if ${method} fails`, async () => {
             const methodSpy = spy();
             methodSpy.mockRejectedValue(new Error(errorText));
-            await expect(deviceCryptoService.trustDevice()).rejects.toThrow(errorText);
+            await expect(deviceTrustCryptoService.trustDevice()).rejects.toThrow(errorText);
           });
 
           test.each([null, undefined])(
@@ -311,7 +311,7 @@ describe("deviceCryptoService", () => {
             async (invalidValue) => {
               const methodSpy = spy();
               methodSpy.mockResolvedValue(invalidValue);
-              await expect(deviceCryptoService.trustDevice()).rejects.toThrow();
+              await expect(deviceTrustCryptoService.trustDevice()).rejects.toThrow();
             }
           );
         }
